+++ packages/lumx-react/src/components/side-navigation/SideNavigation.stories.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+import { mdiAccount } from '@lumx/icons';
 import { Emphasis, SideNavigation, SideNavigationItem } from '@lumx/react';
 
 export default { title: 'LumX components/side-navigation/Side Navigation' };
@@ -27,3 +28,97 @@ export const sideNavigation = () => (
         <SideNavigationItem label="Navigation item" emphasis={Emphasis.low} />
     </SideNavigation>
 );
+
+export const With3Levels = () => {
+    const [l1IsOpen, setL1IsOpen] = React.useState(true);
+    const [l2IsOpen, setL2IsOpen] = React.useState(true);
+    const toggleL1 = () => setL1IsOpen(!l1IsOpen);
+    const toggleL2 = () => setL2IsOpen(!l2IsOpen);
+
+    return (
+        <SideNavigation>
+            <SideNavigationItem label="Level 1" emphasis={Emphasis.high} isOpen={l1IsOpen} onClick={toggleL1}>
+                <SideNavigationItem label="Level 2" emphasis={Emphasis.medium} isOpen={l2IsOpen} onClick={toggleL2}>
+                    <SideNavigationItem label="Level 3" emphasis={Emphasis.low} />
+                </SideNavigationItem>
+            </SideNavigationItem>
+        </SideNavigation>
+    );
+};
+
+export const With3LevelsAndIcons = () => {
+    const [l1IsOpen, setL1IsOpen] = React.useState(true);
+    const [l2IsOpen, setL2IsOpen] = React.useState(true);
+    const toggleL1 = () => setL1IsOpen(!l1IsOpen);
+    const toggleL2 = () => setL2IsOpen(!l2IsOpen);
+
+    return (
+        <SideNavigation>
+            <SideNavigationItem
+                label="Level 1"
+                emphasis={Emphasis.high}
+                isOpen={l1IsOpen}
+                onClick={toggleL1}
+                icon={mdiAccount}
+            >
+                <SideNavigationItem
+                    label="Level 2"
+                    emphasis={Emphasis.medium}
+                    isOpen={l2IsOpen}
+                    onClick={toggleL2}
+                    icon={mdiAccount}
+                >
+                    <SideNavigationItem label="Level 3.1" emphasis={Emphasis.low} isSelected />
+                    <SideNavigationItem label="Level 3.2" emphasis={Emphasis.low} icon={mdiAccount} />
+                    <SideNavigationItem label="Level 3.3" emphasis={Emphasis.low} icon={mdiAccount} />
+                </SideNavigationItem>
+            </SideNavigationItem>
+        </SideNavigation>
+    );
+};
+
+export const With3LevelsAndMultiActions = () => {
+    const [l1IsOpen, setL1IsOpen] = React.useState(true);
+    const [l2IsOpen, setL2IsOpen] = React.useState(true);
+    const toggleL1 = () => setL1IsOpen(!l1IsOpen);
+    const toggleL2 = () => setL2IsOpen(!l2IsOpen);
+    const action3Click = () => alert('Level 3 chevron');
+
+    return (
+        <SideNavigation>
+            <SideNavigationItem
+                label="Level 1"
+                emphasis={Emphasis.high}
+                isOpen={l1IsOpen}
+                // tslint:disable-next-line: ter-no-script-url
+                linkProps={{ href: 'javascript:alert("Level 1")' }}
+                onClick={toggleL1}
+                icon={mdiAccount}
+            >
+                <SideNavigationItem
+                    label="Level 2"
+                    emphasis={Emphasis.medium}
+                    isOpen={l2IsOpen}
+                    // tslint:disable-next-line: ter-no-script-url
+                    linkProps={{ href: 'javascript:alert("Level 2")' }}
+                    onClick={toggleL2}
+                >
+                    <SideNavigationItem
+                        label="Level 3.1"
+                        emphasis={Emphasis.low}
+                        // tslint:disable-next-line: ter-no-script-url
+                        linkProps={{ href: 'javascript:alert("Level 3")' }}
+                        onClick={action3Click}
+                    />
+                    <SideNavigationItem
+                        label="Level 3.2"
+                        emphasis={Emphasis.low}
+                        // tslint:disable-next-line: ter-no-script-url
+                        linkProps={{ href: 'javascript:alert("Level 3")' }}
+                        onClick={action3Click}
+                    />
+                </SideNavigationItem>
+            </SideNavigationItem>
+        </SideNavigation>
+    );
+};